+++ OpenRA.Mods.Cnc/OpenRA.Mods.Cnc.csproj
@@ -70,6 +70,7 @@
     <Compile Include="Activities\HarvesterDockSequence.cs" />
     <Compile Include="CncLoadScreen.cs" />
     <Compile Include="Effects\IonCannon.cs" />
+    <Compile Include="Traits\AttackPopupTurreted.cs" />
     <Compile Include="Traits\Buildings\ProductionAirdrop.cs" />
     <Compile Include="Traits\Buildings\TiberiumRefinery.cs" />
     <Compile Include="Traits\PoisonedByTiberium.cs" />
