+++ OpenRA.Mods.RA/Attack/AttackTurreted.cs
@@ -15,12 +15,12 @@ using OpenRA.Traits;
 namespace OpenRA.Mods.RA
 {
 	[Desc("Actor has a visual turret used to attack.")]
-	class AttackTurretedInfo : AttackFollowInfo, Requires<TurretedInfo>
+	public class AttackTurretedInfo : AttackFollowInfo, Requires<TurretedInfo>
 	{
 		public override object Create(ActorInitializer init) { return new AttackTurreted(init.self, this); }
 	}
 
-	class AttackTurreted : AttackFollow, ITick, ISync
+	public class AttackTurreted : AttackFollow, ITick, ISync
 	{
 		protected IEnumerable<Turreted> turrets;
